这些类的部分属性改为Long型(上次漏了)

+++ 01Project/dolp/02DevelopmentArea/04Code/dolp/src/com/dolplay/dolpbase/system/domain/Organization.java
@@ -24,7 +24,7 @@ public class Organization implements TreeNode {
 	@ColDefine(type = ColType.VARCHAR, width = 500)
 	private String description;
 	@Column
-	private Integer parentOrgId;
+	private Long parentOrgId;
 	@One(target = Organization.class, field = "parentOrgId")
 	private Organization parentOrg;
 	@Many(target = Organization.class, field = "parentOrgId")
@@ -66,11 +66,11 @@ public class Organization implements TreeNode {
 		this.description = description;
 	}
 
-	public Integer getParentOrgId() {
+	public Long getParentOrgId() {
 		return parentOrgId;
 	}
 
-	public void setParentOrgId(Integer parentOrgId) {
+	public void setParentOrgId(Long parentOrgId) {
 		this.parentOrgId = parentOrgId;
 	}
